requirements.txt
altair==4.2.0
asn1crypto==1.5.1
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
DateTime==4.7
decorator==5.1.1
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
jsonschema==4.17.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
numpy==1.23.4
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
plotly==5.11.0
protobuf==3.20.3
PuLP==2.7.0
pyarrow==8.0.0
pycparser==2.21
pycryptodomex==3.16.0
pydeck==0.8.0b4
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
snowflake-connector-python==2.9.0
soupsieve==2.3.2.post1
streamlit==1.14.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
zipp==3.10.0
zope.interface==5.5.1
